+++ src/clj_bugsnag/core.clj
@@ -95,7 +95,8 @@
                  :metaData       (walk/postwalk stringify (merge base-meta meta))}]}))
 
 (defn notify
-  "Post an `exception` to BugSnag.
+  "Notify Bugsnag about the supplied `exception`.
+   By default, this function returns the HTTP response from Bugsnag.
    A second, optional argument may be passed to configure the behavior of the client.
    This map supports the following options.
      - :api-key - The BugSnag API key for your project.
@@ -135,3 +136,40 @@
      (if suppress-bugsnag-response?
        nil
        resp))))
+
+(defn notify-v2!
+  "Notify BugSnag about the supplied `exception`.
+   By default, this function returns nil.
+   A second, optional argument may be passed to configure the behavior of the client, which deviates in behavior from `notify`
+   This map supports the following options.
+     - :api-key - The BugSnag API key for your project.
+                  If this key is missing, the library will attempt to load the Environment variable `BUGSNAG_KEY` and the JVM Property `bugsnagKey` in this order.
+                  If all three values are nil, an exception will be thrown
+     - :project-ns - The BugSnag project name you'd like to report the error to.
+                     Typically the artifact name.
+                     Defaults to \000
+     - :context - The BugSnag 'context' in which an error occurred.
+                  Defaults to nil.
+                  See https://docs.bugsnag.com/platforms/java/other/customizing-error-reports/ for more details
+     - :group - The BugSnag 'group' an error occurred within.
+                Defaults to the exception message for instances of `clojure.lang.ExceptionInfo` or the Class Name of the Exception
+     - :severity - The severity of the error.
+                   Must be one of `info`, `warning`, and `error`.
+                   Defaults to `error`
+     - :user  - A string or map of facets representing the active end user when the error occurred.
+                Defaults to nil
+     - :version - The application version running when the error was reported.
+                  Defaults to the git SHA when possible.
+                  Otherwise nil.
+     - :environment - The deployment context in which the error occurred.
+                      Defaults to `Production`
+     - :meta - A map of arbitrary metadata to associate to the error"
+  ([exception]
+   (notify-v2! exception nil))
+
+  ([exception options]
+   (let [params (exception->json exception options)
+         url    "https://notify.bugsnag.com/"
+         _      (http/post url {:form-params  params
+                                :content-type :json})]
+     nil)))