+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.1.0 / 2021 Oct 21
 
 > This release adds the `:suppress-bugsnag-response?` option to `notify` to make behavior more consistent with other logging interfaces
+> Additionally include a `notify-v2!` which surpresses this output by default
 
 ## v1.0.0 / 2021 Oct 19